Removed context from Backup VM
