+++ app/src/main/kotlin/com/wire/android/util/FileManager.kt
@@ -1,6 +1,7 @@
 package com.wire.android.util
 
 import android.content.Context
+import android.net.Uri
 import com.wire.kalium.logic.data.asset.KaliumFileSystem
 import dagger.hilt.android.qualifiers.ApplicationContext
 import okio.Path
@@ -25,5 +26,17 @@ class FileManager @Inject constructor(@ApplicationContext private val context: C
 
     fun shareWithExternalApp(assetDataPath: Path, assetExtension: String?, onError: () -> Unit) {
         shareAssetFileWithExternalApp(assetDataPath, context, assetExtension, onError)
+    }
+
+    fun copyToTempPath(uri: Uri, tempCachePath: Path): Long {
+        val file = tempCachePath.toFile()
+        var size: Long
+        file.setWritable(true)
+        context.contentResolver.openInputStream(uri).use { inputStream ->
+            file.outputStream().use {
+                size = inputStream?.copyTo(it) ?: -1L
+            }
+        }
+        return size
     }
 }
